Move SinkhornOnline tests
